+++ src/JETLS.jl
@@ -398,8 +398,8 @@ function handle_response_message(
         handle_instantiation_progress_response(server, request_caller)
     elseif request_caller isa AnalysisProgressCaller
         handle_analysis_progress_response(server, request_caller, cancel_flag)
-    elseif request_caller isa ShowDocumentRequestCaller
-        handle_show_document_response(server, msg, request_caller)
+    elseif request_caller isa ShowTextDocumentContentCaller
+        handle_show_text_document_content_response(server, msg, request_caller)
     elseif request_caller isa SetDocumentContentCaller
         handle_apply_workspace_edit_response(server, msg, request_caller)
     elseif request_caller isa DeleteFileCaller
@@ -434,8 +434,6 @@ function handle_response_message(
         handle_profile_progress_response(server, msg, request_caller)
     elseif request_caller isa WorkspaceConfigurationCaller
         handle_workspace_configuration_response(server, msg, request_caller)
-    elseif request_caller isa OpenMacroExpansionCaller
-        handle_open_macro_expansion_response(server, msg, request_caller)
     elseif request_caller isa RegisterCapabilityRequestCaller || request_caller isa UnregisterCapabilityRequestCaller
         # nothing to do
     else